Day20 Part2 semi-automatic
This solution only works for my particular input. It could be generalized to work with any input with the same structure. It might be hard to generalize for any input.

The general approach of finding connected components and then solving them is a good one. For each component, we could determine if that component is stateful or stateless. For stateful components, we would have to understand the cyclic sequence of inputs that would be fed to it in order to understand the cyclic sequence of outputs that would emerge from it.

Another possibility would be to try to analyze each component to try to "understand" its behavior without simulation. That seems very challenging.